CI: Test populate-depot.py in Debian 12 instead of 11
We currently use Debian 12 prereleases to build the actual production
container depots, because Debian 12 has python3-vdf available in the
ordinary package archive, meaning we don't have to bring our own.
Reflect that here.

I've kept this as a template job and an instantiation of that template,
so that when we're eventually thinking about moving CI to Debian 13,
we can test the script in that first, without losing Debian 12 coverage.

Helps: steamrt/tasks#248
Signed-off-by: Simon McVittie <smcv@collabora.com>