pv-wrap: Stop providing /run/host/container-runtime
This was never an intended way to detect the container environment, and
in particular is not provided by non-pressure-vessel containers and
sandboxes such as Flatpak. We should remove it before people start
relying on it. The intended interface is that they should look for
/run/host/container-manager, as per
<https://systemd.io/CONTAINER_INTERFACE/>.

We added this in commit 7460b0cd "pv-wrap: Temporarily provide
/run/host/container-runtime" as a hotfix workaround for a SDL bug, since
updating SDL would have been a riskier change to make to the stable
branch without going through a beta.

Removing this will cause controller hotplug detection to regress in
games that bundled their own copy of SDL between 2.25.x (sometime around
September 2022) and 2.26.4. This regression can be avoided by not bundling
SDL, and instead relying on the Steam Runtime to provide a sufficiently
recent version.

This reverts commit 7460b0cd9fcb1c92a6a97a4a64b52f4c82a671b5.

Signed-off-by: Simon McVittie <smcv@collabora.com>